Removing refactored code
