--- src/app/dynamic-forms/fields/attach-file/attach-file.component.html
@@ -1,33 +1,0 @@
-<div style="width:100%;display:inline" *ngIf="(field.cinchyColumn.dataType == 'Binary')"
-     class="full-width-element divMarginBottom">
-  <div>
-    <div>
-      <div>
-        <fa-icon [icon]="faFile"></fa-icon>
-      </div>
-      &nbsp;
-      <label class="cinchy-label">
-        {{field.label}}
-      </label>
-      <input class='form-control'
-             [disabled]="(field.cinchyColumn.canEdit=== false)"
-             type="file"
-             id="file"
-             (change)="onFileSelected(field,$event)">
-
-      <mat-error *ngIf="showError && (field.cinchyColumn.isMandatory == true &&(field.value =='' || field.value == null))">
-        *{{field.label}} is Required.
-      </mat-error>
-    </div>
-  </div>
-
-  <div>
-    <div *ngIf="fileName" style="margin: 5px;">
-
-      <a class="download-link" (click)="downloadDocument($event, field.value)">
-        Existing file: <strong>{{fileName}}</strong>
-        <mat-icon>cloud_download</mat-icon>
-      </a>
-    </div>
-  </div>
-</div>
