--- src/app/dynamic-forms/fields/checkbox/checkbox.component.html
@@ -1,26 +1,0 @@
-<div *ngIf="(field.cinchyColumn.dataType == 'Yes/No' && field.cinchyColumn.canView)">
-  <fa-icon [icon]="faCheckSquare"></fa-icon>
-  &nbsp;
-  <input class="m-tb-10 mr-5" type='checkbox' [(ngModel)]="field.value" [ngModelOptions]="{standalone: true}"
-         [disabled]="(field.cinchyColumn.canEdit=== false || field.cinchyColumn.isViewOnly || isDisabled)"
-         (change)="valueChanged()" [id]="field.label" />
-  <div class="checkbox-field">
-
-    &nbsp;
-    <label class="pre-formatted" [for]="field.label" [title]="field.caption ? field.caption : ''">
-      {{field.label}}
-      {{field.cinchyColumn.isMandatory == true && (field.value == '' || field.value == null) ? '*' : ''}}
-    </label>
-    <mat-icon *ngIf="field.caption" class="info-icon-checkbox"
-              ngbTooltip="{{field.caption}}"
-              placement="auto"
-              container="body"
-              matTooltipClass="tool-tip-body"
-              matTooltipPosition="above">
-      info
-    </mat-icon>
-  </div>
-  <mat-error *ngIf="showError && (field.cinchyColumn.isMandatory == true &&(field.value =='' || field.value == null))">
-    *{{field.label}} is Required.
-  </mat-error>
-</div>
