--- src/app/dynamic-forms/fields/choice/choice.component.html
@@ -1,44 +1,0 @@
-<!-- Choice field For ShowCase Make it Multi-Select-->
-<div *ngIf=" field.dropdownDataset">
-  <div class="full-width-element divMarginBottom">
-    <div class="link-labels">
-      <div>
-        <fa-icon [icon]="faListUl"></fa-icon>
-      </div>
-      &nbsp;
-      <label class="cinchy-label" [title]="field.caption ? field.caption : ''">
-        {{field.label}}
-        {{field.cinchyColumn.isMandatory == true && (field.value == '' || field.value == null) ? '*' : ''}}
-      </label>
-      <mat-icon *ngIf="field.caption" class="info-icon"
-                ngbTooltip="{{field.caption}}"
-                placement="auto"
-                container="body"
-                matTooltipClass="tool-tip-body"
-                matTooltipPosition="above">
-        info
-      </mat-icon>
-    </div>
-    <mat-form-field class="full-width-element divMarginBottom" *ngIf="field.cinchyColumn.canEdit && !field.cinchyColumn.isViewOnly && !isDisabled">
-      <mat-select class="single-select form-control" [(ngModel)]="field.value"
-                  [disabled]="(field.cinchyColumn.canEdit=== false || field.cinchyColumn.isViewOnly || isDisabled)"
-                  (selectionChange)="callbackEvent(targetTableName, field.cinchyColumn.name, $event, 'value')"
-                  [ngModelOptions]="{standalone: true}">
-        <mat-option [value]=""> </mat-option>
-        <mat-option *ngFor="let dropdownOption of field.dropdownDataset| filterBy: choiceFilter"
-                    [value]="dropdownOption">
-          {{dropdownOption}}
-        </mat-option>
-
-      </mat-select>
-    </mat-form-field>
-    <ng-container *ngIf="!field.cinchyColumn.canEdit || field.cinchyColumn.isViewOnly || isDisabled">
-      <label class="font-14" [innerHTML]="field.value ? field.value : '-'"></label>
-    </ng-container>
-  </div>
-  <!-- isMandatory Validator-->
-  <mat-error class="mat-error-move-up-19"
-             *ngIf="showError && (field.cinchyColumn.isMandatory == true &&(field.value =='' || field.value == null))">
-    *{{field.label}} is Required.
-  </mat-error>
-</div>
