+++ src/app/dynamic-forms/fields/cinchy-my-format.ts
@@ -1,11 +1,11 @@
 export const DisplayFormats = {
-  parse: {
-    dateInput: sessionStorage.getItem('displayFormat'),
-  },
-  display: {
+    parse: {
+      dateInput: sessionStorage.getItem('displayFormat'),
+    },
+    display: {
     dateInput: sessionStorage.getItem('displayFormat'),
     monthYearLabel: 'MMMM YYYY',
     dateA11yLabel: 'DD/MM/YYYY',
     monthYearA11yLabel: 'MMMM YYYY',
-  },
-};
+    },
+  };
