--- src/app/dynamic-forms/fields/link-multichoice/link-multichoice.component.html
@@ -1,147 +1,0 @@
-<form class="dropdown-wrapper">
-  <div class="full-width-element divMarginBottom linkField">
-    <div class="m-b-10">
-      <div class="link-labels">
-        <div>
-          <fa-icon [icon]="field.cinchyColumn.tableId == field.cinchyColumn.LinkTargetTableId ? faSitemap : faShareAlt"></fa-icon>
-        </div>
-        &nbsp;
-        <label class="cinchy-label" [title]="field.caption ? field.caption : ''">
-          {{field.label}}
-          {{field.cinchyColumn.isMandatory == true && (field.value == "" || field.value == null) ? "*" : ""}}
-        </label>
-
-        <mat-icon *ngIf="field.caption && field.cinchyColumn.tableId != field.cinchyColumn.LinkTargetTableId" class="info-icon"
-                  [ngbTooltip]="withcaptiont"
-                  placement="auto"
-                  container="body"
-                  triggers="click"
-                  #t="ngbTooltip"
-                  (mouseenter)="openTooltip(t)"
-                  (mouseleave)="closeTooltip(t)"
-                  matTooltipClass="tool-tip-body"
-                  matTooltipPosition="above">
-          info
-        </mat-icon>
-        <mat-icon *ngIf="!field.caption && field.cinchyColumn.tableId != field.cinchyColumn.LinkTargetTableId" class="info-icon"
-                  [ngbTooltip]="withoutcaptiont"
-                  triggers="click"
-                  placement="auto"
-                  container="body"
-                  #t="ngbTooltip"
-                  (mouseenter)="openTooltip(t)"
-                  (mouseleave)="closeTooltip(t)"
-                  matTooltipClass="tool-tip-body"
-                  matTooltipPosition="above">
-          info
-        </mat-icon>
-        <mat-icon *ngIf="field.caption && field.cinchyColumn.tableId == field.cinchyColumn.LinkTargetTableId" class="info-icon"
-                  [ngbTooltip]="hierarchyWithCaption"
-                  placement="auto"
-                  container="body"
-                  triggers="click"
-                  #t="ngbTooltip"
-                  (mouseenter)="openTooltip(t)"
-                  (mouseleave)="closeTooltip(t)"
-                  matTooltipClass="tool-tip-body"
-                  matTooltipPosition="above">
-          info
-        </mat-icon>
-        <mat-icon *ngIf="!field.caption && field.cinchyColumn.tableId == field.cinchyColumn.LinkTargetTableId" class="info-icon"
-                  [ngbTooltip]="hierarchyWithoutCaption"
-                  placement="auto"
-                  container="body"
-                  triggers="click"
-                  #t="ngbTooltip"
-                  (mouseenter)="openTooltip(t)"
-                  (mouseleave)="closeTooltip(t)"
-                  matTooltipClass="tool-tip-body"
-                  matTooltipPosition="above">
-          info
-        </mat-icon>
-        <ng-template #withcaptiont>
-          {{this.field.caption}}  <br /> <br /> From the <b> {{this.field.cinchyColumn.linkTargetColumnName}} </b> field in the <a [href]="tableSourceURL" target="_blank">  {{this.field.cinchyColumn.linkTargetTableName}}  </a> table.
-        </ng-template>
-        <ng-template #withoutcaptiont>
-          From the <b> {{this.field.cinchyColumn.linkTargetColumnName}} </b> field in the <a [href]="tableSourceURL" target="_blank">  {{this.field.cinchyColumn.linkTargetTableName}}  </a> table.
-        </ng-template>
-        <ng-template #hierarchyWithCaption>
-          {{this.field.caption}}  <br /> <br /> From the <b> {{this.field.cinchyColumn.linkTargetColumnName}} </b> field in the {{this.field.cinchyColumn.linkTargetTableName}} table.
-        </ng-template>
-        <ng-template #hierarchyWithoutCaption>
-          From the <b> {{this.field.cinchyColumn.linkTargetColumnName}} </b> field in the {{this.field.cinchyColumn.linkTargetTableName}} table.
-        </ng-template>
-
-      </div>
-      <mat-select class="form-control" multiple #multiSelect [formControl]="myControl" *ngIf="!isLoading && !downloadLink"
-                  disableOptionCentering>
-        <ng-container *ngIf="dropdownSetOptions">
-          <mat-option>
-            <ngx-mat-select-search class="multi-search"
-                                   [showToggleAllCheckbox]="true" (toggleAll)="toggleSelectAll($event)"
-                                   [formControl]="multiFilterCtrl"
-                                   placeholderLabel="Search..."
-                                   noEntriesFoundLabel="no matching bank found">
-            </ngx-mat-select-search>
-          </mat-option>
-
-          <ng-container *ngIf="filteredListMulti | async; let list">
-            <ng-container *ngIf="list.length < maxLimitForMaterialSelect; else cdkOptions">
-              <mat-option class="all-options" *ngFor="let dropdownOption of list"
-                          [value]="dropdownOption"
-                          [title]="dropdownOption.label"
-                          (click)="optionClicked(dropdownOption)">
-                {{dropdownOption.label}}
-              </mat-option>
-            </ng-container>
-
-            <ng-template #cdkOptions>
-              <cdk-virtual-scroll-viewport [itemSize]="48" [style.height.px]=4*48 [minBufferPx]="300"
-                                           [maxBufferPx]="600">
-                <mat-option class="all-options" *cdkVirtualFor="let dropdownOption of filteredListMulti | async"
-                            [value]="dropdownOption"
-                            [title]="dropdownOption.label"
-                            (click)="optionClicked(dropdownOption)">
-                  {{dropdownOption.label}}
-                </mat-option>
-              </cdk-virtual-scroll-viewport>
-            </ng-template>
-          </ng-container>
-        </ng-container>
-      </mat-select>
-      <!--        <ng-container>
-              <div *ngIf="dropdownSetOptions && !isLoading">
-                <angular2-multiselect [data]="dropdownSetOptions" [formControl]="myControl"
-                                      [(ngModel)]="selectedValues"
-                                      [settings]="dropdownSettings"
-                                      (onSelect)="onItemSelect($event)"
-                                      (onDeSelect)="OnItemDeSelect($event)"
-                                      (onSelectAll)="onSelectAll($event)"
-                                      (onDeSelectAll)="onDeSelectAll($event)">
-                </angular2-multiselect>
-              </div>
-              </ng-container>-->
-      <mat-error *ngIf="showError && (field.cinchyColumn.isMandatory == true &&(field.value =='' || field.value == null))">
-        *{{field.label}} is Required.
-      </mat-error>
-
-      <ng-container *ngIf="downloadLink">
-        <div *ngFor="let item of downloadableLinks" style="margin-top: 6px">
-          <a *ngIf="downloadLink" [href]="item.fileUrl" [title]="'Download ' + item.fileName">
-            <img *ngIf="renderImageFiles && fileNameIsImage(item.fileName)" style="height: 100px" [src]="item.fileUrl" />
-            <span [style.marginLeft.px]="renderImageFiles && fileNameIsImage(item.fileName) ? 12 : 0">{{item.fileName}}</span>
-          </a>
-          <mat-icon class="file-delete-icon" (click)="onDeleteFile(item)" title="Delete">close</mat-icon>
-        </div>
-        <input #fileInput multiple class='form-control'
-               *ngIf="!(field.cinchyColumn.canEdit=== false)"
-               type="file"
-               (change)="onFileSelected($event)"
-               [style.marginTop.px]="downloadableLinks ? 14 : 0">
-      </ng-container>
-    </div>
-  </div>
-</form>
-<div *ngIf="isLoading" class="is-loading">
-  <mat-spinner diameter="20"></mat-spinner>
-</div>
