+++ src/app/dynamic-forms/fields/rich-text/rich-text.component.html
@@ -93,4 +93,7 @@
   <ng-container *ngIf="field.cinchyColumn.canEdit === false || field.cinchyColumn.isViewOnly || isDisabled">
     <label class="pre-formatted">{{field.value | json}}</label>
   </ng-container>
+
+
+  <label class="pre-formatted">{{field.value | json}}</label>
 </div>
