--- src/app/dynamic-forms/fields/textbox/textbox.component.html
@@ -1,50 +1,0 @@
-<div *ngIf="(field.cinchyColumn.dataType == 'Text' && field.cinchyColumn.textColumnMaxLength <= 500
-  && field.cinchyColumn.canView)"
-     class="full-width-element divMarginBottom">
-  <div class="link-labels">
-    <div>
-      <fa-icon [icon]="faAlignLeft"></fa-icon>
-    </div>
-    &nbsp;
-    <label class="cinchy-label" [title]="field.caption ? field.caption : ''">
-      {{field.label}}
-      {{field.cinchyColumn.isMandatory == true && (field.value == '' || field.value == null) ? '*' : ''}}
-    </label>
-    <mat-icon *ngIf="field.caption" class="info-icon"
-              ngbTooltip="{{field.caption}}"
-              placement="auto"
-              container="body"
-              matTooltipClass="tool-tip-body"
-              matTooltipPosition="above">
-      info
-    </mat-icon>
-  </div>
-  <ng-container *ngIf="!field.cinchyColumn.isViewOnly && !isDisabled && field.cinchyColumn.canEdit && showActualField">
-    <input class='form-control'
-           [disabled]="(field.cinchyColumn.canEdit=== false || field.cinchyColumn.isViewOnly  || field.label=== 'File Name' || isDisabled)"
-           type="text" [(ngModel)]="field.value"
-           (blur)="callbackEvent(targetTableName, field.cinchyColumn.name, $event, 'value')"
-           [ngModelOptions]="{standalone: true}">
-    <mat-error *ngIf="showError && (field.cinchyColumn.isMandatory == true &&(field.value =='' || field.value == null))">
-      *{{field.label}} is Required.
-    </mat-error>
-  </ng-container>
-
-  <label class="pre-formatted"
-         *ngIf="!field.cinchyColumn.canEdit || field.cinchyColumn.isViewOnly  || field.label=== 'File Name' || isDisabled"
-         [innerHTML]="field.value || '-'"></label>
-
-  <ng-container *ngIf="showImage">
-    <img [ngClass]="{
-        'cinchy-images': size === '${ImageType.default}' || size === '${ImageType.medium}',
-        'cinchy-images-large' : size === '${ImageType.large}',
-        'cinchy-images-small' : size === '${ImageType.small}'
-      }" *ngIf="field.value" [src]="field.value">
-    <p *ngIf="!field.value">-</p>
-  </ng-container>
-
-  <ng-container *ngIf="showLinkUrl">
-    <a *ngIf="field.value" [href]="field.value">Open</a>
-    <label *ngIf="!field.value">-</label>
-  </ng-container>
-</div>
